Remove title from proposal comments
When reading a proposal's comments with a screen reader, it reads out
the entire timestamp of a comment, like `2019-01-02 20:33:56 -0800`.
When read out loud, this takes a long time, is disruptive to the reading
experience, and sounds a bit silly.

Further, this timestamp is nowhere on the page and doesn't seem to serve
a purpose.

Instead, this commit removes the `title` field on each proposal comment.
This way, when entering the comment section, a user hears the "X
[Public/Internal] Comments" header, then is able to dive directly into
each comment itself. This content flow is much smoother, making for a
better user experience.