Redo proposal abstract length validation
The abstract length was originally 600 characters. Carriage return
characters would typically be included in the standard length
validation, so the limit was increased to 1000. This commit creates a
custom validation to strip out the carriage return characters and count
the resulting string length. If _THAT_ is greater than 600 characters
then there is an error.