Allow updating tags after proposal is finalized
Tags can be a helpful way to organize talks into buckets as they are
being evaluated and accepted. They can also be useful for categorizing
talks when creating the schedule to make sure you have a good balance
across days and rooms. This allows tags to be updated after a talk has
been finalized.