Add PaperTrail gem, update gems
Add PaperTrail gem to allow for highlighting diffs in proposals in the
future. I thought it would be helpful to at least start collecting the
data. I also updated gems, including making the versions less specific
in the Gemfile. The most notable update is that Capybara went from 2.x
to 3.x, which required some changes in the specs.