Inital pass at a dynamic Programs page
This was first implemented on the flagrant/website_spike branch

it creates a new route at 'website/program'
that calls a new programs controller, where it picks up an events live program_sessions.
It then displays those program sessions dynamically using a _session partial

Additionally it adds 2 scopes on the program_session model that enable filtering between talks (regular_sessions) and workshops
Right now this is done with hard coded strings, that query the name of Session_format table.
Will update that soon to something that is less brittle.