Initial pass at a dynamic Schedule Page
This was first implemented on the flagrant/website_spike branch

it creates a new route at 'website/schedule'
that calls a new schedule controller, which populates a schedule in the same fashion that the CFP application is currently using to export a JSON schedule.
It then displays the schedule sessions dynamically using session partials

Additionally it uses a helper method to sort/group the program sessions into time blocks.
Currently disabled links leading to program page, and the day sub navigation.