Access event throught current_website
For the most part the idea is to access the event data, by going through the website.
This will prevent a current_website and current_event from getting out of step.

Details
Move current_website to the application controller
Enforce website requirment on the program controller
Obtain event data throught the current_website when in program views
(Temp) link to program page in default template