Schedule page sub navigation
handling page sub-navigation with stimulus.
Brief description of what that looks like this.

A stimulus controller warps around the schedule.
Each day of the schedule has a unique id. ex:'schedule-day-one'
The controler is given an inital element to display, labeled as 'displayedId'.
Each day of the schedule is lableled as 'hideable' using a stimulus data target.

When the page loads, the contoller loops through the 'hideable' targets. And it hides them.
excepts the the target with the id that matches displayedId.
When a link in the sub nav is clicked, a controller action updates the display.

Simple as that.
We will be able to set the displayedId dynamically, i.e. 'the schedule opens to the correct page'.