reuse time_slot_decorator to display schedule blocks
Quick refactor. Objective was to remove the TimeSlotSerializer from the schedule controller.
It was being used to 'decorate' the JSON file that was being sent to the conference websites.
We don't need that JSON data anymore.

By simply adding two helper methods (.room, and .shortened_display_description) we can totally reuse
the time_slot decerator to decorate the schedule.
Hey. That's pretty neat.

A final refactoring goal would be to eliminate the singleton method in the WebsiteScheduleHelper and
Write a standalone model Schedule, that specifically quired the information it wanted directly from ActiveRecord.
Moving the work one step closer to the database.

Gonna get some tests on my backside before trying that though.