website sponsor page
add a route, controller, a slew of partials, and a little bit of CSS to dynamicailly display an events sponsors on the event website.

Implementation
- the sponsors controller has a before action I labeled :load_sponsors
- popluates two instance variables @sponsors, and @sponsors_by_tier using the events 'published' sponsors.
- loop through data, display with partials

I set it up in this way so that we could reuse the :load_sponsors method on different website controllers (i.e. PagesController which will need to display the Sponsors)

Additional Notes
- I hit a .variant error while trying to resize an uploaded .svg, will need to consider image types when adding preview to the sponsor _form
- Im still missing some data attributes that need to be added to the SponsorModel. offer_url, offer_text, offer_headline, etc... but we are fleshing it out.