Create a website Schedule Page (#269)
* Initial pass at a dynamic Schedule Page

This was first implemented on the flagrant/website_spike branch

it creates a new route at 'website/schedule'
that calls a new schedule controller, which populates a schedule in the same fashion that the CFP application is currently using to export a JSON schedule.
It then displays the schedule sessions dynamically using session partials

Additionally it uses a helper method to sort/group the program sessions into time blocks.
Currently disabled links leading to program page, and the day sub navigation.

* use different naming for schedule sessions

moved away from session, for obvious reasons

* Schedule page sub navigation

handling page sub-navigation with stimulus.
Brief description of what that looks like this.

A stimulus controller warps around the schedule.
Each day of the schedule has a unique id. ex:'schedule-day-one'
The controler is given an inital element to display, labeled as 'displayedId'.
Each day of the schedule is lableled as 'hideable' using a stimulus data target.

When the page loads, the contoller loops through the 'hideable' targets. And it hides them.
excepts the the target with the id that matches displayedId.
When a link in the sub nav is clicked, a controller action updates the display.

Simple as that.
We will be able to set the displayedId dynamically, i.e. 'the schedule opens to the correct page'.

* reuse time_slot_decorator to display schedule blocks

Quick refactor. Objective was to remove the TimeSlotSerializer from the schedule controller.
It was being used to 'decorate' the JSON file that was being sent to the conference websites.
We don't need that JSON data anymore.

By simply adding two helper methods (.room, and .shortened_display_description) we can totally reuse
the time_slot decerator to decorate the schedule.
Hey. That's pretty neat.

A final refactoring goal would be to eliminate the singleton method in the WebsiteScheduleHelper and
Write a standalone model Schedule, that specifically quired the information it wanted directly from ActiveRecord.
Moving the work one step closer to the database.

Gonna get some tests on my backside before trying that though.

* feature testing schedule page

tests are as follows
- update time_slot and verify the change is reflected on the schedule
- update program session, and verify the change is reflected on the schedule
- destory a time slot, and verify that it is no longer displayed on the schedule
- test the sub nav links, verify that sessions are displayed on the correct date
  - this one needs some work.

* clean up haml syntax