dynamic sponsors footer componnet
this PR uses stimulus to generate sponsors footer element.

Steps to create component
in the page body create a div with the following data attributes
 - data-controller="sponsors-footer"
 - data-sponsors-footer-event-slug-value="<: ENTER EVENT SLUG >"

Thats it.

this will mount a stimulus controller to the page.
That controller will issue a fetch request to the sponsors controller, and return with the sponsors_footer partial.
That partial will then be inserted into the stimulus contoller element.