sponsor banner ad stimulus controller
For refrence this feature is being inmplemented using the same method as the sponsor footer component
Enables a website creator/editor to add a dynamic sponsor banner ad to any stataic page.

The command that will be required will be
```
<div data-controller="banner-ads" data-banner-ads-event-slug-value="railsconf-2022" data-banner-ads-index-value="0"></div>
```

For now I coded in an inital index, will probably change this be a random index in the ads array.