schedule fly out
this could use some additional work to make it a bit cleaner.
But for now it gets the job done.

fly out partial is hidden outside the right side of the screen.
When a title is clicked, the elemnt matching that titles id is 'revealed' as the side panel slides out.