Display placeholder message if filter produces no resutls
When filtering, if an applied filter resluts in all session content being hidden the stimulus controller will apply an empty class to the container.
That will lean in to some CSS magic to dispalys a "No results message" to the user

Also add some hover effect to the filter icon.