collapse schedule block when cards filtered out
Had to resort to something interesting here, and Can redo this if someone suggests a better method.
In order to collapse the schedule blocks when their time slots had been filtered out I had to create custom class that defined a filtercontainer in a way that it could have sub containers (in this case 'schedule-block-containers').

Now when the stimulus controller calls apply filter it sill loop over the containers and update their apperence. IF a container has sub-containers, it will loop through them and update their apperance as well (i.e. give them an 'empty' class. Essentially collapse them)