Update schedule page - Program/Schedule page filter
* allow time slots to be associated with an event sponsor

this will enable the ability for a time slot to be displayed on an event website with an associated event sponsor

* finalize schedule appearence

redesign schedule to align with Gina's design.
Next steps is wiring up the schedule fly out.

* schedule fly out

this could use some additional work to make it a bit cleaner.
But for now it gets the job done.

fly out partial is hidden outside the right side of the screen.
When a title is clicked, the elemnt matching that titles id is 'revealed' as the side panel slides out.

* program filter pane

use stimulus to filter the content on the program page.
method used:

- filter partial is added to the page side panel
- form in filter partial, has a filter stimulus controller wrapped around it
- when form inputs are changed, stimulus fires event to update page contents
- update page contents using dom_ids for relevent filterable attributes

* add filter to schedule side panel

Reuse the filter controller by initilizing it with the class name of the 'cards' that it is filtering through
Also reverse the direction that the filter works. All values start unchecked, checking box, filters view & then updates counts

Additional Notes
- Rename schedule card partial to time_slot
- Make filtering an async function, update counts after completion
- Hide elements using 'display:none' instead of hidden attribute (hidden wouldn't work with the CSS Grid on the schedule)

* fixes per PR code review

* update schedule test to consider session format configuration

* schedule loads displaying the event current day

* Display placeholder message if filter produces no resutls

When filtering, if an applied filter resluts in all session content being hidden the stimulus controller will apply an empty class to the container.
That will lean in to some CSS magic to dispalys a "No results message" to the user

Also add some hover effect to the filter icon.

* add sponsor presents to time slot display

additionally I moved the sub_nav style to shared, so program and schedule were using the same sub nav css
Also added a check for empty containers on filter reset.

* collapse schedule block when cards filtered out

Had to resort to something interesting here, and Can redo this if someone suggests a better method.
In order to collapse the schedule blocks when their time slots had been filtered out I had to create custom class that defined a filtercontainer in a way that it could have sub containers (in this case 'schedule-block-containers').

Now when the stimulus controller calls apply filter it sill loop over the containers and update their apperence. IF a container has sub-containers, it will loop through them and update their apperance as well (i.e. give them an 'empty' class. Essentially collapse them)

* pull the bindings out of my schedule spec

move time helper to rails_helper file

* rework 'schedule load on correct date feature' with caching consideration

Use on page JS to load the schedule on the correct page.

When the page loads, the schedule will display a link for every event day.
The JS will determine what day it is for the user, and
- IF it is before the event OR 1 day after event it will load on the first event day
- IF it is during the event it will loop over the event days, and click on the current day.