Add field to support Texas vs RI speaker question (#327)
This is a quick and dirty solution that we will want to remove as soon
as RubyConf 2022 is done. We need to ask speakers where they feel most
comfortable giving their talk, and the extra field that the CFP actually
supports is open ended rather than being constrained to options.